Move Callback interface to seperate package
